+++ components/ld2415h/number/min_speed_threshold_number.h
@@ -11,7 +11,7 @@ class MinSpeedThresholdNumber : public number::Number, public Parented<LD2415HCo
   MinSpeedThresholdNumber() = default;
 
  protected:
-  void control(uint8_t speed) override;
+  void control(float speed) override;
 };
 
 }  // namespace ld2415h
