+++ components/ld2415h/number/relay_trigger_duration_number.cpp
@@ -3,7 +3,7 @@
 namespace esphome {
 namespace ld2415h {
 
-void RelayTriggerDurationNumber::control(uint8_t duration) {
+void RelayTriggerDurationNumber::control(float duration) {
   this->publish_state(duration);
   this->parent_->set_relay_trigger_duration(duration);
 }
