+++ components/ld2415h/number/relay_trigger_speed_number.cpp
@@ -3,7 +3,7 @@
 namespace esphome {
 namespace ld2415h {
 
-void RelayTriggerSpeedNumber::control(uint8_t speed) {
+void RelayTriggerSpeedNumber::control(float speed) {
   this->publish_state(speed);
   this->parent_->set_relay_trigger_speed(speed);
 }
