+++ components/ld2415h/number/vibration_correction_number.cpp
@@ -3,7 +3,7 @@
 namespace esphome {
 namespace ld2415h {
 
-void VibrationCorrectionNumber::control(uint8_t correction) {
+void VibrationCorrectionNumber::control(float correction) {
   this->publish_state(correction);
   this->parent_->set_vibration_correction(correction);
 }
